Removed unnecessary lines.

+++ Assets/Scripts/Player/PlayerBullet/PlayerBulletModel.cs
@@ -1,11 +1,7 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
 public class PlayerBulletModel
 {
     public float bulletSpeed;
-    public float bulletDamage;
+    public int bulletDamage;
    public PlayerBulletModel()
     {
         bulletSpeed = 250f;
